backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-generativeai==0.8.3
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
